pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend", "backend.api", "backend.db", "backend.schemas", "backend.services", "backend.utils"]

[project]
name = "loqui-tts"
version = "1.0.0"
description = "Loqui - A beautiful local TTS application powered by Chatterbox"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "chatterbox-tts @ git+https://github.com/resemble-ai/chatterbox.git",
    "pyloudnorm>=0.2.0",
    "torchaudio>=2.0.0",
    "pydantic>=2.0.0",
    "aiofiles>=24.1.0",
    "greenlet>=3.0.0",
    "setuptools>=68.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
